==> src/analyses/Brecht/01_selection_and_matrix_construction.ipynb <==
# Discard unrelated articles and create matrix
This notebook uses the functions in `preprocessing_functions.py` to select only potentially relevant articles. And then uses them to create a co-occurrence matrix for each language.  
**Note: This notebook can take several hours to run.**

## Inputs
This notebook requires inputs not in the repository (for storage reasons). Namely the extracted wikidumps. The notebook `citynet/src/analyses/1wikidump Downloading and Parsing.ipynb` can be used to download these and also contains instructions for extracting them. The path to the directory above the one where these extracted documents are saved should be added to `extr_dir`. The actual folder where these are stored should be named according to the following pattern `f"{lang}wiki_extracted"` for the notebook to work as presented. Otherwise change the `wiki_dir` variable.

## Outputs
Selection of potentially relevant articles  
`../../../../data/enwiki/enwikidump_msXX.csv
../../../../data/enwiki/frwikidump_msXX.csv`

Co-occurrence matrices  
`../../output/en_matrix.csv
../../output/fr_matrix.csv`

City-pair co-occurrences  
`../../citypairs_en.csv
../../citypairs_fr.csv`



## Packages and loading files

# load packages
import re, string
import pandas as pd
import numpy as np
import os
import time
from tqdm.notebook import tqdm
import unidecode

# import from script
from preprocessing_functions import *

data_dir = '../../../../data/' # directory where selected articles will be saved, change if you want to save these elsewhere
out_dir = '../../output/'
in_dir = '../../input/'
extr_dir = # path/to/wikidump/extracted

# list of languages to iterate over
lang_list = ['en', 'fr']

# cities dataframe
fp = '../../input/List_of_cities_300k.csv' # path to csv with city information
cities = pd.read_csv(fp, sep=';')

## Discard unrelated articles
This creates multiple .csv files with the 'id', 'title' and 'text' of Wikipedia articles with at least 2 toponym occurrences. 

for lang in lang_list:
    wiki_dir = f"{lang}wiki_extracted" # language specific directory
    
    base_dir = os.path.join(extr_dir, wiki_dir) 

    # load city list
    name_col = f'Mua_{lang}' 
    cities_ls = [unidecode.unidecode(city_component) 
             for city in cities[name_col] 
             for city_component in city.split('-')]
    
    # do the whole preprocessing/extraction thing
    preprocess(base_dir, data_dir, lang, cities_ls, overwrite_protection = False)

## Create Co-occurence matrix
Iterates over the .csv files created in the previous chunk to identify toponym co-occurences and add these to a matrix. Function documentation within `preprocessing_functions.py`

# create co-occurence matrix
for lang in lang_list:
    name_col = f'Mua_{lang}'

    # list of complete city names
    city_l = [unidecode.unidecode(city) for city in cities[name_col]]
    
    # folder of extracted streams
    inputfp = os.path.join(data_dir, f'{lang}wiki/')
    
    #- loop over .csv files create dataframes
    df = pd.DataFrame(columns = ['article_id', 'title', 'text'])
    # iterate over directory for each file path, create a dataframe
    for file in os.listdir(inputfp):
        fp = os.path.join(inputfp, file)
        df_temp = pd.read_csv(fp)
        df = pd.concat([df, df_temp])
    
    # create matrix
    matrix = process_corpus(df.text, city_l)
    
    # save matrix
    FILENAME = f"{lang}_matrix.csv"

    write_matrix(matrix = matrix, 
                 outdir = out_dir, 
                 filename = FILENAME)


print(time.asctime(time.localtime(time.time())))


# create city-pairs dataframe

for lang in lang_list:
    # load matrix
    FILENAME = f"{lang}_matrix.csv"
    matrix = pd.read_csv(os.path.join(out_dir, FILENAME))
    matrix.set_index('index', inplace=True)
    
    
    # create city pairs list
    citylink = create_citylink(matrix)
    
    # as dataframe
    df = pd.DataFrame(citylink, columns = ['city_A', 'city_B', f'col_{lang}'])
    
    # save as csv
    FILENAME = f"citypairs_{lang}.csv"
    fp = os.path.join(out_dir, FILENAME)
    
    df.to_csv(fp, index = False)

df.head()
