==> src/analyses/Brecht/03_match_coordinates.ipynb <==
# Creation of dataframe for gravity model

Join list of cities with a dataset that contains co-ordinates to allow for mapping and distance calculation. 
This notebook also assigns city-level dummies, such as whether or not a city is in a certain region, or whether English or French is spoken there.

## Required Input
This notebook requires input not available in this repository, namely `euro-global-map-shp/data/FullEurope/BuiltupP.shp`. This dataset can be obtained from [https://www.mapsforeurope.org/datasets/euro-global-map](https://www.mapsforeurope.org/datasets/euro-global-map). 

Furthermore it requires QGIS (or equivalent) for the addition of one of the datapoints.

Other inputs are available within the repository

## Generated output

`../../input/maps/nodes.shp`

import pandas as pd
import geopandas as gpd
import numpy as np
import os
import re
import unidecode
from collections import Counter

INDIR = "../../input"
OUTDIR = "../../output"
DATADIR = "../../../../data" # path/to/non/repo/files
FILE_cities = "List_of_cities_300k.csv"
FILE_coords = "euro-global-map-shp/data/FullEurope/BuiltupP.shp" #path/to/eucoordinates/shapefile

# load citylist
list_of_cities = pd.read_csv(os.path.join(INDIR, FILE_cities), sep=";")
list_of_cities.head()

## Select relevant coordinates

# load data with city coordinates
built_up = gpd.read_file(os.path.join(DATADIR, FILE_coords))

built_up.head()

`BuiltUpP.shp` contains 4 different name variables. `NAMN1`, `NAMN2`, `NAMA1`, and `NAMA2`. The first two are names in up to two national languages. The latter are transliterated versions of those names into ASCII characters. 

muas = list(list_of_cities.Mua)

# create a dictionary of toponyms and their country codes
mua_dict = dict()

for index, row in list_of_cities.iterrows():
    if row['Code_Country'] == 'UK': 
        mua_dict[row['Mua']] = 'GB'
    else: 
        mua_dict[row['Mua']] = row['Code_Country']
    

## First Match
Match the cities in the built_up dataset with the cities in the CITYNET network by confirming that name and country match.

cities = []
for index, row in built_up.iterrows(): 
    if row['NAMA1'] in mua_dict and row['ICC'] == mua_dict[row['NAMA1']]:
        cities.append(row['NAMA1'])
    elif row['NAMN1'] in mua_dict and row['ICC'] == mua_dict[row['NAMN1']]:
        cities.append(row['NAMN1'])
    elif row['NAMN2'] in mua_dict and row['ICC'] == mua_dict[row['NAMN2']]: 
        cities.append(row['NAMN2'])

# check how many cities have been matched
len(cities)

# check if any cities have been matched more than once
city_counts = Counter(cities) 
for key in city_counts:
    if city_counts[key] > 1:
        print(key)

# cities not automatically matched
leftovers = []
for city in muas:
    if city not in cities:
        leftovers.append(city)

print(leftovers)

## Outcome
Matched: 135 cities  
False Match: Bremen and Leeds duplicates

Cities not matched:  
'Brussels', 'Kobenhavn', 'Lyon', 'Sofia', 'Essen-Oberhausen', 'Marseille', 'Antwerp', 'Bochum-Herne', 'Gelsenkirchen-Bottrop', 'Belfast', 'Palma de Mallorca', 'Wuppertal', 'Castellammare di Stabia-Torre Annunziata', 'Plovdiv', 'Alicanta', 'La Coruna'

## Causes for failure to match

These cities remain unmatched for a variety of issues.
* language differences (e.g. Brussels, Alicante)
* different name (e.g. Marseille, Mallorca, A Coruna)
* different transliteration (e.g.Plovdiv, Kobenhavn, Sofia)
* MUA is made up of multiple cities (e.g. Castellammare di Stabia-Torre Annunziata)
* Belfast isn't matched cause the country codes don't match (the coordinate dataset uses a separate CC for northern ireland)

One of the cities has no match because it is not present in the coordinate dataset: Wuppertal

Considering the small size of the unmatched city list. A dictionary was created manually to match the diverging names in the two datasets

## Second Match
Create a dictionary for variants, and split double names
### Create dictionary

# dictionary of non matching names
manual_names_v = {"Pl?vdiv":"Plovdiv", 
                "S?fiya":"Sofia", 
                "Koebenhavn": "Kobenhavn", 
                "Brussel": "Brussels", 
                "Antwerpen": "Antwerp",
                "A Coruna": "La Coruna", 
                "Alacant/Alicante": "Alicanta", 
                "Lyon 1er Arrondissement": "Lyon", 
                "Marseille 1er Arrondissement": "Marseille", 
                "Mallorca": "Palma de Mallorca"}

# add first one of the compound cities to the dictionary
for city in leftovers:
    if '-' in city:
        manual_names_v[city.split('-')[0]] = city

# full dictionary
manual_names_v

# reverse of dictionary
manual_names_r = {value: key for key, value in manual_names_v.items()}

# add new spellings to the original city : country code dictionary
for key in manual_names_v:
    # create new mua_dict key from the name dictionary and gives it the same country code as the corresponding name
    mua_dict[key] = mua_dict[manual_names_v[key]]

### Match with this new dictionary

cities = []
for index, row in built_up.iterrows(): 
    if row['NAMA1'] in mua_dict and row['ICC'] == mua_dict[row['NAMA1']]:
        cities.append(tuple(row))
    elif row['NAMN1'] in mua_dict and row['ICC'] == mua_dict[row['NAMN1']]:
        cities.append(tuple(row))
    elif row['NAMN2'] in mua_dict and row['ICC'] == mua_dict[row['NAMN2']]: 
        cities.append(tuple(row))
    # don't require country code match for Belfast (there is only one Belfast in the dataset)
    elif row['NAMA1'] == "Belfast": 
        cities.append(tuple(row))

df = pd.DataFrame(cities, columns = built_up.columns)
df.shape

df.head()

# exclude the duplicate Bremen and Leeds
# Wrong BREMEN 27573
# Wrong LEEDS 63132
duplicate_IDs = [27573, 63132]

# done this way because that way the projection metadata is preserved
gdf = built_up[(~built_up.OBJECTID.isin(duplicate_IDs)) & (built_up.OBJECTID.isin(df.OBJECTID))]

# save as shapefile (outside this repo)
fp = os.path.join(DATADIR, "city_coordinates.shp")
gdf.to_file(fp)

### Manual Addition
Wuppertal is not in the EGM dataset, and was added manually in QGIS (3.16.12 Hannover) based on OpenStreetMap data.

- Loaded the just created `city_coordinates.shp` in QGIS . 
- Used OSM plugin and overlaid it. Toggled editing then `edit` > `add point feature`
- Added point object to city_coordinates layer. And added 'Wuppertal' for 'NAMN1', and 'DE' for 'ICC'.
- Next `save layer edits` and untoggled editting.
- Finally right-clicked city_coordinates layer > `Export` > `Save Features as` > "city_coordinates_complete.shp"

# load file with wuppertal coordinates also included
coords = gpd.read_file(os.path.join(DATADIR, "city_coordinates_complete.shp"))

coords.crs

coords.head()

## Create Cities dataframe

following columns: Mua, Mua_en, Mua_fr, population, country_code, geometry, (dummies: fr_dum, en_dum, additional disambiguation dummies).  



df = list_of_cities.copy()
df = df[['id_MUA', 'Mua', 'Mua_en', 'Mua_fr', 'SizeMUA1k', 'Code_Country']]
df.head()

# change the names in coords so that they match the names in the cities dataframe
coords.replace({'NAMA1': manual_names_v}, inplace=True)

output_df = []
for index, row in coords.iterrows(): 
    for index2, row2 in df.iterrows(): 
        if row['NAMA1'] == row2['Mua']:
            new_row = list(row2[['id_MUA', 'Mua', 'Mua_en', 'Mua_fr', 'SizeMUA1k', 'Code_Country']])
            new_row.append(row.geometry)
            output_df.append(tuple(new_row))
        elif row['NAMN1'] == row2['Mua']:
            new_row = list(row2[['id_MUA', 'Mua', 'Mua_en', 'Mua_fr', 'SizeMUA1k', 'Code_Country']])
            new_row.append(row.geometry)
            output_df.append(tuple(new_row))
        elif row['NAMN2'] == row2['Mua']: 
            new_row = list(row2[['id_MUA', 'Mua', 'Mua_en', 'Mua_fr', 'SizeMUA1k', 'Code_Country']])
            new_row.append(row.geometry)
            output_df.append(tuple(new_row))

# transform to GeoDataFrame
output_df = gpd.GeoDataFrame(output_df, columns = ['id_MUA', 'Mua', 'Mua_en', 'Mua_fr', 'SizeMUA1k', 'Code_Country', 'geometry'], crs = coords.crs)

output_df.plot()

# rename columns to avoid name being cut off when saving
output_df.rename(columns = {'Code_Country': 'CC', 'SizeMUA1k': 'POP', 'Mua_en': 'MUA'}, inplace = True)
output_df.drop(['Mua_fr', 'Mua', 'id_MUA'], axis = 1, inplace = True)

output_df.head()

## if you want to save a version of the coordinate shapefile without the dummy values
# fp = os.path.join(INDIR, "maps/city_coordinates.shp")
# output_df.to_file(fp)

### European area zone dummies
*  Central and Eastern Europe
*  Northern Europe
*  Southern Europe
*  Western Europe  

Categorisation derived from EUROVOC:  
[https://eur-lex.europa.eu/browse/eurovoc.html?params=72,7206,914#arrow_914](https://eur-lex.europa.eu/browse/eurovoc.html?params=72,7206,914#arrow_914)

Only countries included in our dataset are listed here, of course more countries belong to these categories. 


| **Central / East** | **West**         | **North** | **South** |
|--------------------|------------------|-----------|-----------|
| Bulgaria           | Austria          | Denmark   | Greece    |
| Czech Republic     | Belgium          | Estonia   | Italy     |
| Hungary            | France           | Finland   | Malta     |
| Poland             | Germany          | Latvia    | Portugal  |
| Romania            | Ireland          | Lithuania | Spain     |
| Slovakia           | Netherlands      | Norway    |       -   |
|       -            | Switzerland      | Sweden    |       -   |
|       -            | (United Kingdom) |     -     |       -   |



set(output_df.CC)

west = ['AT', 'BE', 'FR', 'DE', 'IE', 'NL', 'CH', 'UK']
south =  ['GR', 'IT', 'MT', 'PT', 'ES']
north = ['DK', 'EE', 'FI', 'LV', 'LT', 'NO', 'SE']
central_east = ['PL', 'RO', 'HU', 'CZ', 'BG', 'SK']

len(west) + len(south) + len(north) + len(central_east)

output_df['WEST'] = np.where((output_df['CC'].isin(west)), 1, 0)
output_df['SOUTH'] = np.where((output_df['CC'].isin(south)), 1, 0)
output_df['NORTH'] = np.where((output_df['CC'].isin(north)), 1, 0)
output_df['CEAST'] = np.where((output_df['CC'].isin(central_east)), 1, 0)

def region_lab(row): 
    if row['WEST'] == 1: 
        region = 'west'
    elif row['CEAST'] == 1:
        region = 'central_east'
    elif row['NORTH'] == 1: 
        region = 'north'
    else: 
        region = 'south'
    return region

output_df['REGION'] = output_df.apply(region_lab, axis = 1)

### Language Sphere dummies

Add language sphere dummies (1 if the city is English or French speaking).

# create english language sphere dummy based on country code
output_df['EN_DUM'] = np.where((output_df['CC']== 'UK') | (output_df['CC'] == 'IE'), 1, 0)

# French-speaking cities in countries which speak french only in certain regions
french_sphere = ['Charleroi', 'Liège', 'Brussels', 'Geneva']

# french dummy CC French or city from previous list
fr_dum = []
for i, row in output_df.iterrows(): 
    if row['MUA'] in french_sphere:
        fr_dum.append(1)
    elif row['CC'] == 'FR':
        fr_dum.append(1)
    else:
        fr_dum.append(0)

# add to dataframe
output_df['FR_DUM'] = fr_dum

fp = os.path.join(INDIR, "maps/nodes.shp")
# output_df.to_file(fp)
