==> src/analyses/Brecht/05_gravity_model.Rmd <==
---
title: "gravity_model"
author: "Brecht Nijman"
date: "6/3/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Load Data & Packages

```{r libraries}
library(tidyverse)
library(ggplot2)
library(stargazer)
library(magrittr)
```

```{r}
# knitr::root.dir("")
```

```{r}
getwd()
```


```{r load edges}
setwd("/Users/brecht/Documents/citynet_2")
edges <- read_delim("citynet/src/output/edges.csv",
                         ";", escape_double = FALSE, 
                         col_types = 
                           cols(border_DUM = col_factor(levels = c("0", "1")), 
                                FR_con = col_factor(levels = c("0", "1")), 
                                FR_int = col_factor(levels = c("0", "1")), 
                                EN_con = col_factor(levels = c("0", "1")), 
                                EN_int = col_factor(levels = c("0", "1")),
                                CEAST = col_factor(levels = c("0", "1")), 
                                WEST = col_factor(levels = c("0", "1")), 
                                NORTH = col_factor(levels = c("0", "1")), 
                                SOUTH = col_factor(levels = c("0", "1"))), 
                         trim_ws = TRUE)

edges <- tibble(edges)


edges_nz <- read_delim("citynet/src/output/edges_nz.csv",
                         ";", escape_double = FALSE, 
                         col_types = 
                           cols(border_DUM = col_factor(levels = c("0", "1")), 
                                FR_con = col_factor(levels = c("0", "1")), 
                                FR_int = col_factor(levels = c("0", "1")), 
                                EN_con = col_factor(levels = c("0", "1")), 
                                EN_int = col_factor(levels = c("0", "1")),
                                CEAST = col_factor(levels = c("0", "1")), 
                                WEST = col_factor(levels = c("0", "1")), 
                                NORTH = col_factor(levels = c("0", "1")), 
                                SOUTH = col_factor(levels = c("0", "1"))), 
                         trim_ws = TRUE)

edges <- tibble(edges)
head(edges_nz)
```


## Gravity Model

```{r}
# # subset without 0s
# edges_nz_en <- subset(edges, col_en > 0)
# edges_nz_fr <- subset(edges, col_fr > 0)

# subset only cities with population of over a million
edges_mil <-
  edges_nz %>% 
  filter(POP_A > 999 & POP_B > 999)

head(edges_mil)
```

### French

```{r}
# base model french
lm.fr1 <- lm(log(col_fr) ~ log(POP_A) + log(POP_B) + log(distance), edges_nz)
summary(lm.fr1)
```

```{r}
# fr model with border
lm.fr2 <- lm(log(col_fr) ~ log(POP_A) + log(POP_B) + log(distance) + border_DUM, edges_nz)
summary(lm.fr2)
```

```{r}

# fr model with border + language sphere only french
lm.fr3 <- lm(log(col_fr) ~ log(POP_A) + log(POP_B) + log(distance) + border_DUM + FR_con, edges_nz)
summary(lm.fr3)

```


```{r}

# fr model with border + language sphere only french
lm.fr4 <- lm(log(col_fr) ~ log(POP_A) + log(POP_B) + log(distance) + border_DUM + FR_con + EN_con, edges_nz)
summary(lm.fr4)

```

```{r}
# fr model 1 mil base
lm.fr5 <- lm(log(col_fr) ~ log(POP_A) + log(POP_B) + log(distance), edges_mil)
summary(lm.fr5)
```

```{r}
# fr model 1 mil full
lm.fr6 <- lm(log(col_fr) ~ log(POP_A) + log(POP_B) + log(distance) + border_DUM + FR_con + EN_con, edges_nz)
summary(lm.fr6)
```


### English

```{r}
# base model english
lm.en1 <- lm(log(col_en) ~ log(POP_A) + log(POP_B) + log(distance), edges_nz)
summary(lm.en1)
```

```{r}
# en model with border
lm.en2 <- lm(log(col_en) ~ log(POP_A) + log(POP_B) + log(distance) + border_DUM, edges_nz)
summary(lm.en2)
```

```{r}

# fr model with border + language sphere only french
lm.en3 <- lm(log(col_en) ~ log(POP_A) + log(POP_B) + log(distance) + border_DUM + EN_con, edges_nz)
summary(lm.en3)

```


```{r}

# fr model with border + language sphere only french
lm.en4 <- lm(log(col_en) ~ log(POP_A) + log(POP_B) + log(distance) + border_DUM + FR_con + EN_con, edges_nz)
summary(lm.en4)

```


```{r}
# en model 1 mil base
lm.en5 <- lm(log(col_en) ~ log(POP_A) + log(POP_B) + log(distance), edges_mil)
summary(lm.en5)
```

```{r}
# en model 1 mil full
lm.en6 <- lm(log(col_en) ~ log(POP_A) + log(POP_B) + log(distance) + border_DUM + FR_con + EN_con, edges_nz)
summary(lm.en6)
```

### Regional effects

```{r}
# fr with regional dummies
lm.fr.cc1 <- lm(log(col_fr) ~ log(POP_A) + log(POP_B) + log(distance) + CEAST + WEST + NORTH + SOUTH, edges_nz)
summary(lm.fr.cc1)
```

```{r}
# fr with regional dummies
lm.fr.cc2 <- lm(log(col_fr) ~ log(POP_A) + log(POP_B) + log(distance) + CEAST + WEST + NORTH + SOUTH + border_DUM, edges_nz)
summary(lm.fr.cc2)
```

```{r}
# en with regional dummies
lm.en.cc1 <- lm(log(col_en) ~ log(POP_A) + log(POP_B) + log(distance) + CEAST + WEST + NORTH + SOUTH, edges_nz)
summary(lm.en.cc1)
```

```{r}
# with regional dummies and border
lm.en.cc2 <- lm(log(col_en) ~ log(POP_A) + log(POP_B) + log(distance) + CEAST + WEST + NORTH + SOUTH + border_DUM, edges_nz)
summary(lm.en.cc2)
```


### Add predictions

```{r}
# add english residuals (base model with border)
edges_nz$res_en2 <- lm.en2$residuals
edges_nz$res_fr2 <- lm.fr2$residuals
```



```{r}
edges_nz$pred_en2 <- lm.en2$fitted.values
edges_nz$pred_fr2 <- lm.fr2$fitted.values
```


```{r}
# calculate as percentage
setwd("/Users/brecht/Documents/citynet_2")

write_delim(edges_nz,"citynet/src/output/edges_nz.csv", delim = ';')
```
