==> src/analyses/Brecht/06_nodes.ipynb <==
# Nodes

**This notebook should be run after the R Markdown file `05_gravity_model.Rmd`**

Calculate the degree of the nodes (i.e. how often a single city co-occurs with any other) in each corpus.  
Additionally calculate the predicted ln(degree) and residuals of each node.

At the end there is one edge related section where the predictions & residuals are added to the edge shapefile to allow for mapping for the final report.

## Inputs

## Outputs
nodes. 
`../../output/maps/nodes.shp`

`../../output/nodes.csv`

`../../output/maps/edges_nz.shp`

# packages
import pandas as pd
import geopandas as gpd
import numpy as np
import os
import unidecode

## Functions

def degree_count(df, en_column, fr_column):
    '''creates a dictionary for english and french occurrences of each city which is updated with 
    the counts from the appropriate column
    '''
    
    en_degrees = dict()
    for city in cities: 
        en_degrees[city] = 0
    
    fr_degrees = en_degrees.copy()
    
    for i, row in df.iterrows():
        en_degrees[row.city_A] += row[en_column]
        en_degrees[row.city_B] += row[en_column]
        fr_degrees[row.city_A] += row[fr_column]
        fr_degrees[row.city_B] += row[fr_column]
    
    return en_degrees, fr_degrees

## Load File

# directories
INDIR = "../../input"
OUTDIR = "../../output"
DATADIR = "../../../../data"
FILE_cities = "List_of_cities_300k.csv"

# load list of cities
list_of_cities = pd.read_csv(os.path.join(INDIR, FILE_cities), sep=';')
cities = [unidecode.unidecode(city) for city in list_of_cities.Mua_en]
df_cities = pd.DataFrame(cities, columns = ["city"])

# load nodes shapefile
fp = os.path.join(INDIR, "maps/nodes.shp")
nodes = gpd.read_file(fp)

# load the edges shapefile
fp = os.path.join(OUTDIR, "maps/edges_nz.shp")
edges_gdf = gpd.read_file(fp)

# load edges with predictions
# fp = os.path.join(OUTDIR, "edges_nz.csv"),
edges = pd.read_csv("../../output/edges_nz.csv", sep=';')

edges.head()

### Degrees (occurrences of a node)

# calculate occurences in french and english
en_degrees, fr_degrees = degree_count(edges, 'col_en', 'col_fr')


# add calculated occurences to dataframe
df_cities['en_degree'] = en_degrees.values()
df_cities['fr_degree'] = fr_degrees.values()

df_cities.sort_values('en_degree', ascending=False).head(10)

# calculate proportional occurences in french and english
en_degrees, fr_degrees = degree_count(edges, 'col_prop_en', 'col_prop_fr')


# add calculated occurences to dataframe
df_cities['en_deg_prop'] = en_degrees.values()
df_cities['fr_deg_prop'] = fr_degrees.values()

df_cities.head()

## 

# calculate predicted ln(occurences) in french and english
en_degrees, fr_degrees = degree_count(edges, 'pred_en2', 'pred_fr2')


# add calculated predicted occurences to dataframe
df_cities['fr_pred'] = fr_degrees.values()
df_cities['en_pred'] = en_degrees.values()

# total residuals of cities in french and english
en_degrees, fr_degrees = degree_count(edges, 'res_en2', 'res_fr2')


# add total residuals to dataframe
df_cities['fr_res'] = fr_degrees.values()
df_cities['en_res'] = en_degrees.values()

# percentage residual of prediction
def estimation(row, pred_col, res_col): 
    percentage = (row[res_col] / row[pred_col]) * 100
    
    return round(percentage, 2)

df_cities['fr_estim'] = df_cities.apply(estimation, pred_col = 'fr_pred', res_col = 'fr_res', axis = 1)
df_cities['en_estim'] = df_cities.apply(estimation, pred_col = 'en_pred', res_col = 'en_res', axis = 1)

df_cities[['city','en_pred','fr_pred','en_res','fr_res','fr_estim', 'en_estim']].sort_values('fr_estim', ascending=True).head(10)

## Merge with Coordinates for mapping

nodes.head()

nodes['MUA'] = [unidecode.unidecode(city) for city in nodes['MUA']]

nodes = pd.merge(nodes, df_cities, how='left', left_on = 'MUA', right_on = 'city')

nodes.drop(['city'], axis = 1, inplace = True)

nodes.head()

# save to .shp file
fp = os.path.join(OUTDIR, "maps/nodes.shp")
# nodes.to_file(fp)

# drop geometry
nodes.drop(['geometry'], axis = 1, inplace = True)

# save as csv
fp = os.path.join(OUTDIR, "nodes.csv")
nodes.to_csv(fp, index = False, sep=';') 

nodes[['MUA', 'fr_degree', 'en_degree', 'fr_pred', 'en_pred']].sort_values('en_pred', ascending=False).head(10)

## Edges
adding predicted values for mapping in QGIS

edges

edges[['city_A', 'city_B']].iloc[8052] == edges_gdf[['city_A', 'city_B']].iloc[8052]

# add predictions and residuals to the edges GeoDataFrame
edges_gdf[['res_en2', 'res_fr2', 'pred_en2', 'pred_fr2']] = edges[['res_en2', 'res_fr2', 'pred_en2', 'pred_fr2']]

# save edges geodataframe
fp = os.path.join(OUTDIR, 'maps/edges_nz2.shp')
# edges_gdf.to_file(fp)
